+++ antpack/__init__.py
@@ -1,9 +1,9 @@
 #Used for determining the version if running setup.
 #Only change if building a new version.
-__version__ = "0.1.5"
+__version__ = "0.2.0"
 
-from .consensus_update_tools import build_consensus_alignment
-from .single_chain_annotator import SingleChainAnnotator
-from .sequence_scoring_tool import SequenceScoringTool
-from .humanization_tool import HumanizationTool
-from .utilities.data_compile_utils import get_position_dict
+from .numbering_tools import SingleChainAnnotator, MultiChainAnnotator
+from .vj_tools.vj_gene_assignment import VJGeneTool
+from .scoring_tools import SequenceScoringTool, HumanizationTool, LiabilitySearchTool
+from .numbering_tools.consensus_update_tools import build_consensus_alignment
+from .utilities.vj_gene_database_writer import update_vj_databases